requirements.txt
requests
dotenv
lxml
beautifulsoup4
Pillow